TOTS-78 Fixed issue with TOTS layers not loading when refreshing TODS.
